Fix issues raised by static code analysis.
This continues the work in 0d4e625358aa64a575e93b9b62dd0fdd7404071e by
fixing places where we copy structures when looping qmaps.

It also removes an unused variable which somehow QtCreator did not warn
about.

It also fixes a mismatch between signal definitions and how it was used
in a connect statement.

+++ Engine/Engine.cpp
@@ -32,7 +32,7 @@ void Engine::prepare() {
 }
 
 void Engine::dump() {
-    const double elapsed = timer->elapsed() / 1000;
+    const double elapsed = double(timer->elapsed()) / 1000;
 
     QMap<QString, int>::const_iterator i = processed_events_of_type.constBegin();
     while (i != processed_events_of_type.constEnd()) {
